Revert "Update on enabled (#1145)"
This reverts commit f30fe5e75edbadf2c5b81b5097253a210037c294.
