+++ api_app/analyzers_manager/apps.py
@@ -11,4 +11,9 @@ class AnalyzersManagerConfig(AppConfig):
     name = "api_app.analyzers_manager"
 
     def ready(self):
+        from .serializers import AnalyzerConfigSerializer  # to avoid import issue
+
         set_permissions(settings.DEFAULT_CACHE)
+        # we "greedy cache" the config at start of application
+        # because it is an expensive operation
+        AnalyzerConfigSerializer.read_and_verify_config()
