+++ api_app/analyzers_manager/views.py
@@ -1,5 +1,6 @@
 # This file is a part of IntelOwl https://github.com/intelowlproject/IntelOwl
 # See the file 'LICENSE' for copying permission.
+
 import logging
 
 from drf_spectacular.utils import extend_schema as add_docs
