+++ api_app/connectors_manager/apps.py
@@ -6,3 +6,10 @@ from django.apps import AppConfig
 
 class ConnectorsManagerConfig(AppConfig):
     name = "api_app.connectors_manager"
+
+    def ready(self):
+        from .serializers import ConnectorConfigSerializer  # to avoid import issue
+
+        # we "greedy cache" the config at start of application
+        # because it is an expensive operation
+        ConnectorConfigSerializer.read_and_verify_config()
