+++ api_app/connectors_manager/dataclasses.py
@@ -10,17 +10,12 @@ from .serializers import ConnectorConfigSerializer
 
 __all__ = ["ConnectorConfig"]
 
-from ..models import PluginCredential
-
 
 @dataclasses.dataclass
 class ConnectorConfig(AbstractConfig):
     maximum_tlp: str
 
     serializer_class = ConnectorConfigSerializer
-
-    def _get_type(self) -> str:
-        return PluginCredential.PluginType.ANALYZER
 
     def get_full_import_path(self) -> str:
         return f"api_app.connectors_manager.connectors.{self.python_module}"
