+++ api_app/core/dataclasses.py
@@ -50,10 +50,6 @@ class AbstractConfig:
     config: _Config
     verification: _SecretsVerification
 
-    @abstractmethod
-    def _get_type(self) -> str:
-        raise NotImplementedError()
-
     def __post_init__(self):
         secrets_values = list(self.secrets.values())
         params_values = list(self.params.values())
@@ -101,12 +97,9 @@ class AbstractConfig:
             }
         else:
             _filtered_secrets = self.secrets
-        for key_name in _filtered_secrets.keys():
+        for key_name, secret in _filtered_secrets.items():
             secrets[key_name] = secrets_store.get_secret(
-                key_name,
-                default=None,
-                plugin_type=self._get_type(),
-                plugin_name=self.name,
+                secret.env_var_key, default=None
             )
 
         return secrets
