+++ api_app/models.py
@@ -284,29 +284,3 @@ class CustomConfig(models.Model):
                         plugin["params"][param]["value"] = custom_configs[
                             plugin["name"]
                         ][param]
-
-
-class PluginCredential(models.Model):
-    class PluginType(models.TextChoices):
-        ANALYZER = "1", "Analyzer"
-        CONNECTOR = "2", "Connector"
-
-    type = models.CharField(choices=PluginType.choices, max_length=2)
-    attribute = models.CharField(max_length=128)
-    value = models.TextField()
-    updated_at = models.DateTimeField(auto_now=True)
-    plugin_name = models.CharField(max_length=128)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["type", "attribute", "plugin_name"],
-                name="unique_plugin_credential_entry",
-            )
-        ]
-
-        indexes = [
-            models.Index(
-                fields=["type", "plugin_name", "attribute"],
-            ),
-        ]
