+++ api_app/views.py
@@ -17,7 +17,6 @@ from rest_framework import serializers as rfs
 from rest_framework import status, viewsets
 from rest_framework.decorators import action, api_view
 from rest_framework.exceptions import ValidationError
-from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
 
@@ -32,7 +31,7 @@ from .analyzers_manager import controller as analyzers_controller
 from .analyzers_manager.constants import ObservableTypes
 from .filters import JobFilter
 from .helpers import get_now
-from .models import TLP, CustomConfig, Job, PluginCredential, Status, Tag
+from .models import TLP, CustomConfig, Job, Status, Tag
 from .serializers import (
     AnalysisResponseSerializer,
     CustomConfigSerializer,
@@ -41,7 +40,6 @@ from .serializers import (
     JobListSerializer,
     JobSerializer,
     ObservableAnalysisSerializer,
-    PluginCredentialSerializer,
     TagSerializer,
     multi_result_enveloper,
 )
@@ -670,14 +668,3 @@ class CustomConfigViewSet(viewsets.ModelViewSet):
             request._full_data = request.data.copy()
         request.data["owner"] = request.user.id
         return super().update(request, *args, **kwargs)
-
-
-class PluginCredentialViewSet(viewsets.ModelViewSet):
-    queryset = PluginCredential.objects.all()
-    serializer_class = PluginCredentialSerializer
-    pagination_class = None
-    permission_classes = (
-        (IsAuthenticated,)
-        if settings.NON_ADMIN_CAN_EDIT_PLUGIN_SECRETS
-        else (IsAdminUser,)
-    )
