+++ docker/env_file_app_ci
@@ -11,10 +11,6 @@ OLD_JOBS_RETENTION_DAYS=2
 INTELOWL_WEB_CLIENT_DOMAIN=localhost
 # Storage
 LOCAL_STORAGE=True
-
-# Allow non-admin users to edit plugin secrets from GUI
-NON_ADMIN_CAN_EDIT_PLUGIN_SECRETS=False
-
 # OAuth2
 GOOGLE_CLIENT_ID=test_intelowl.apps.googleusercontent.com
 GOOGLE_CLIENT_SECRET=the_supreme_secret
