+++ docker/env_file_app_template
@@ -12,10 +12,6 @@ OLD_JOBS_RETENTION_DAYS=3
 INTELOWL_WEB_CLIENT_DOMAIN=localhost
 # Storage
 LOCAL_STORAGE=True
-
-# Allow non-admin users to edit plugin secrets from GUI
-NON_ADMIN_CAN_EDIT_PLUGIN_SECRETS=False
-
 # OAuth2
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
@@ -38,6 +34,72 @@ ELASTICSEARCH_HOST=
 # consult to: https://django-elasticsearch-dsl.readthedocs.io/en/latest/settings.html
 ELASTICSEARCH_NO_OF_SHARDS=1
 ELASTICSEARCH_NO_OF_REPLICAS=0
+
+# analyzers secrets (optional)
+CUCKOO_API_KEY=
+SHODAN_KEY=
+ABUSEIPDB_KEY=
+CAPESANDBOX_KEY=
+ANOMALI_THREATSTREAM_API_USER=
+ANOMALI_THREATSTREAM_API_KEY=
+AUTH0_KEY=
+SECURITYTRAILS_KEY=
+GSF_KEY=
+OTX_KEY=
+# Format: user|pass
+CIRCL_CREDENTIALS=
+VT_KEY=
+VT_INTELLIGENCE_KEY=
+HA_KEY=
+INTEZER_KEY=
+INQUEST_API_KEY=
+DOCGUARD_KEY=
+FIRST_MISP_API=
+MISP_KEY=
+MAXMIND_KEY=
+DNSDB_KEY=
+HONEYDB_API_ID=
+HONEYDB_API_KEY=
+HUNTER_API_KEY=
+CENSYS_API_ID=
+CENSYS_API_SECRET=
+ONYPHE_KEY=
+GREYNOISE_API_KEY=
+PULSEDIVE_API_KEY=
+INTELX_API_KEY=
+WHOISXMLAPI_KEY=
+UNPAC_ME_API_KEY=
+URLSCAN_API_KEY=
+PHISHTANK_API_KEY=
+IPINFO_KEY=
+ZOOMEYE_KEY=
+TRIAGE_KEY=
+EMAILREP_KEY=
+MWDB_KEY=
+WIGLE_KEY=
+XFORCE_KEY=
+XFORCE_PASSWORD=
+SSAPINET_KEY=
+GOOGLE_APPLICATION_CREDENTIALS=/opt/deploy/intel_owl/configuration/service_account_keyfile.json
+MALPEDIA_KEY=
+GREEDYBEAR_API_KEY=
+MALPEDIA_TOKEN=
+YARAIFY_KEY=
+# DEHASHED_AUTH_KEY format is `email:api-key`
+DEHASHED_AUTH_KEY=
+OPENCTI_KEY=
+YETI_KEY=
+SPYSE_API_KEY=
+DRAGONFLY_API_KEY=
+DRAGONFLY_API_URL=https://dragonfly.certego.net
+BITCOINABUSE_KEY=
+VIRUSHEE_API_KEY=
+STALKPHISH_KEY=
+
+# connectors secrets (optional)
+CONNECTOR_MISP_KEY=
+CONNECTOR_OPENCTI_KEY=
+CONNECTOR_YETI_KEY=
 
 # Test tokens
 TEST_JOB_ID=1
