+++ docs/source/Installation.md
@@ -100,14 +100,6 @@ In the `env_file_app`, configure different variables as explained below.
 * `DJANGO_SECRET`: random 50 chars key, must be unique. If you do not provide one, Intel Owl will automatically set a new secret on every run.
 * `INTELOWL_WEB_CLIENT_DOMAIN` (example: `localhost`/`mywebsite.com`): the web domain of your instance, this is used for generating links to analysis results.
 
-
-### Deprecated environment configuration
-The following variables are deprecated and will be removed in the future.
-The new way to configure plugin secrets is to use the `Plugin Secrets` page in the GUI.
-If you had previously specified any variables in the environment, run `docker exec -ti intelowl_uwsgi python3 manage.py migrate_secrets`
-and then remove those secrets from the env file.
-
-```text
 **Optional** variables needed to enable specific analyzers:
 * `ABUSEIPDB_KEY`: AbuseIPDB API key
 * `AUTH0_KEY`: Auth0 API Key
@@ -163,7 +155,7 @@ and then remove those secrets from the env file.
 
 **Advanced** additional configuration:
 * `OLD_JOBS_RETENTION_DAYS`: Database retention for analysis results (default: 3 days). Change this if you want to keep your old analysis longer in the database.
-```
+
 ### Database configuration (required)
 In the `env_file_postgres`, configure different variables as explained below.
