+++ frontend/src/components/Routes.jsx
@@ -15,7 +15,6 @@ const PluginsContainer = React.lazy(() => import("./plugins/PluginsContainer"));
 const Dashboard = React.lazy(() => import("./dashboard/Dashboard"));
 const ScanForm = React.lazy(() => import("./scan/ScanForm"));
 const UserConfig = React.lazy(() => import("./me/config/UserConfig"));
-const Secrets = React.lazy(() => import("./me/secrets/Secrets"));
 /*
 lazy imports to enable code splitting
 */
@@ -72,19 +71,10 @@ const authRoutesLazy = [
   },
   /* CustomConfig */
   {
-    path: "/me/config",
+    path: "/me/config/*",
     element: (
       <Suspense fallback={<FallBackLoading />}>
         <UserConfig />
-      </Suspense>
-    ),
-  },
-  /* Secrets */
-  {
-    path: "/me/plugin-secrets",
-    element: (
-      <Suspense fallback={<FallBackLoading />}>
-        <Secrets />
       </Suspense>
     ),
   },
