+++ frontend/src/components/me/organization/MyOrgPage.jsx
@@ -21,7 +21,6 @@ export default function MyOrgPage() {
     error: respErr,
     organization,
     fetchAll,
-    noOrg,
   } = useOrganizationStore(
     React.useCallback(
       (state) => ({
@@ -29,7 +28,6 @@ export default function MyOrgPage() {
         error: state.error,
         organization: state.organization,
         fetchAll: state.fetchAll,
-        noOrg: state.noOrg,
       }),
       []
     )
@@ -37,10 +35,10 @@ export default function MyOrgPage() {
 
   // on component mount
   React.useEffect(() => {
-    if (Object.keys(organization).length === 0 && !noOrg) {
+    if (Object.keys(organization).length === 0) {
       fetchAll();
     }
-  }, [organization, fetchAll, noOrg]);
+  }, [organization, fetchAll]);
 
   // page title
   useTitle(
