+++ frontend/src/components/me/organization/OrgConfig.jsx
@@ -17,7 +17,6 @@ export default function OrgConfig() {
     organization,
     fetchAll,
     isUserOwner,
-    noOrg,
   } = useOrganizationStore(
     React.useCallback(
       (state) => ({
@@ -26,7 +25,6 @@ export default function OrgConfig() {
         organization: state.organization,
         fetchAll: state.fetchAll,
         isUserOwner: state.isUserOwner,
-        noOrg: state.noOrg,
       }),
       []
     )
@@ -34,10 +32,10 @@ export default function OrgConfig() {
 
   // on component mount
   React.useEffect(() => {
-    if (Object.keys(organization).length === 0 && !noOrg) {
+    if (Object.keys(organization).length === 0) {
       fetchAll();
     }
-  }, [noOrg, organization, fetchAll]);
+  }, [organization, fetchAll]);
 
   // page title
   useTitle(
@@ -47,29 +45,28 @@ export default function OrgConfig() {
     { restoreOnUnmount: true }
   );
 
+  const NewOrg = (
+    <Alert color="secondary" className="mt-3 mx-auto">
+      <section>
+        <h5 className="text-warning text-center">
+          You are not owner of any organization.
+        </h5>
+        <p className="text-center">
+          You can choose to create a new organization.
+        </p>
+      </section>
+      <section className="text-center">
+        <OrgCreateButton onCreate={fetchAll} />
+      </section>
+    </Alert>
+  );
+
   return (
     <LoadingBoundary
       loading={loading}
       error={respErr}
       render={() => {
-        if (!isUserOwner)
-          return (
-            <Container>
-              <Alert color="secondary" className="mt-3 mx-auto">
-                <section>
-                  <h5 className="text-warning text-center">
-                    You are not owner of any organization.
-                  </h5>
-                  <p className="text-center">
-                    You can choose to create a new organization.
-                  </p>
-                </section>
-                <section className="text-center">
-                  <OrgCreateButton onCreate={fetchAll} />
-                </section>
-              </Alert>
-            </Container>
-          );
+        if (!isUserOwner) return NewOrg;
         return (
           <Container>
             <h4>{organization.name}&apos;s custom configuration</h4>
@@ -85,21 +82,7 @@ export default function OrgConfig() {
       renderError={({ error }) => (
         <Row>
           {error?.response?.status === 404 ? (
-            <Container>
-              <Alert color="secondary" className="mt-3 mx-auto">
-                <section>
-                  <h5 className="text-warning text-center">
-                    You are not owner of any organization.
-                  </h5>
-                  <p className="text-center">
-                    You can choose to create a new organization.
-                  </p>
-                </section>
-                <section className="text-center">
-                  <OrgCreateButton onCreate={fetchAll} />
-                </section>
-              </Alert>
-            </Container>
+            <NewOrg />
           ) : (
             <ErrorAlert error={error} />
           )}
