+++ intel_owl/settings/__init__.py
@@ -65,7 +65,3 @@ CELERY_QUEUES = secrets.get_secret("CELERY_QUEUES", "default").split(",")
 # AWS
 AWS_SECRETS = secrets.get_secret("AWS_SECRETS", False) == "True"
 AWS_SQS = secrets.get_secret("AWS_SQS", False) == "True"
-
-NON_ADMIN_CAN_EDIT_PLUGIN_SECRETS = (
-    secrets.get_secret("NON_ADMIN_CAN_EDIT_PLUGIN_SECRETS", False) == "True"
-)
