+++ intel_owl/tasks.py
@@ -3,7 +3,7 @@
 
 from __future__ import absolute_import, unicode_literals
 
-from celery import shared_task, signals
+from celery import shared_task
 
 from api_app import crons
 from api_app.analyzers_manager import controller as analyzers_controller
@@ -79,19 +79,3 @@ def start_connectors(
     connectors_controller.start_connectors(
         job_id, connectors_to_execute, runtime_configuration
     )
-
-
-@shared_task()
-def build_config_cache(*args, **kwargs):
-    from api_app.analyzers_manager.serializers import AnalyzerConfigSerializer
-    from api_app.connectors_manager.serializers import ConnectorConfigSerializer
-
-    # we "greedy cache" the config at start of application
-    # because it is an expensive operation
-    AnalyzerConfigSerializer.read_and_verify_config()
-    ConnectorConfigSerializer.read_and_verify_config()
-
-
-@signals.worker_ready.connect
-def worker_ready_connect(*args, **kwargs):
-    build_config_cache()
